fix loss, tidy

+++ constants.py
@@ -1,4 +1,4 @@
 import torch
 
 
-DTYPE = torch.bfloat16
+DTYPE = torch.float64  # change to bfloat16 for speed, but accuracy is better for debugging
